services/motion/camera.py
"""
Camera Motion Tracking using CoTracker

Tracks camera motion through video frames to extract:
- Pan (horizontal movement)
- Tilt (vertical movement)
- Zoom (in/out)
- Roll (rotation)
- Shake (handheld camera effect)

Uses CoTracker for dense point tracking, then derives camera motion.
"""

import asyncio
import logging
from pathlib import Path
from typing import Any, Optional

import numpy as np

from .extractor import CameraMotion

logger = logging.getLogger(__name__)


class CameraTracker:
    """
    Tracks camera motion through video using point tracking.

    Uses CoTracker v2 for dense point tracking, then estimates
    camera motion from point trajectories.

    Usage:
        tracker = CameraTracker()
        camera_motion = await tracker.track(video_path)

        for cm in camera_motion:
            print(f"Frame {cm.frame_number}: pan={cm.pan:.2f}, zoom={cm.zoom:.2f}")
    """

    model_version: str = "cotracker-2.0"

    def __init__(
        self,
        config: Optional[Any] = None,
        grid_size: int = 20,  # Point grid density
        sample_rate: int = 1,  # Track every Nth frame
    ):
        """
        Initialize camera tracker.

        Args:
            config: Configuration object
            grid_size: Size of point grid for tracking
            sample_rate: Process every Nth frame
        """
        self.config = config
        self.grid_size = grid_size
        self.sample_rate = sample_rate

        self._model = None
        self._device = "cuda" if self._check_gpu() else "cpu"

    def _check_gpu(self) -> bool:
        """Check if GPU is available."""
        try:
            import torch
            return torch.cuda.is_available()
        except ImportError:
            return False

    async def _load_model(self):
        """Load the CoTracker model."""
        if self._model is not None:
            return

        logger.info(f"Loading CoTracker on {self._device}")

        # TODO: Implement actual CoTracker loading
        # This would load the CoTracker model from Facebook Research

        self._model = "placeholder"
        logger.info("CoTracker loaded")

    async def track(
        self,
        video_path: Path,
        start_frame: int = 0,
        end_frame: Optional[int] = None,
    ) -> list[CameraMotion]:
        """
        Track camera motion through video.

        Args:
            video_path: Path to video file
            start_frame: Starting frame
            end_frame: Ending frame (None = all)

        Returns:
            List of CameraMotion per frame
        """
        await self._load_model()

        logger.info(f"Tracking camera motion in {video_path}")

        # Get video properties
        fps, frame_count = await self._get_video_info(video_path)

        if end_frame is None:
            end_frame = frame_count

        # Track points through video
        point_tracks = await self._track_points(video_path, start_frame, end_frame)

        # Estimate camera motion from point tracks
        camera_motion = await self._estimate_camera_motion(
            point_tracks, fps, start_frame, end_frame
        )

        logger.info(f"Tracked camera motion for {len(camera_motion)} frames")
        return camera_motion

    async def _get_video_info(self, video_path: Path) -> tuple[float, int]:
        """Get video FPS and frame count."""
        import subprocess
        import json

        cmd = [
            "ffprobe",
            "-v", "quiet",
            "-print_format", "json",
            "-show_streams",
            str(video_path),
        ]

        try:
            result = subprocess.run(cmd, capture_output=True, text=True, timeout=30)
            data = json.loads(result.stdout)

            video_stream = next(
                (s for s in data.get("streams", []) if s.get("codec_type") == "video"),
                {},
            )

            fps = eval(video_stream.get("r_frame_rate", "30/1"))
            frame_count = int(video_stream.get("nb_frames", 0))

            return fps, frame_count

        except Exception as e:
            logger.error(f"Failed to get video info: {e}")
            return 30.0, 0

    async def _track_points(
        self,
        video_path: Path,
        start_frame: int,
        end_frame: int,
    ) -> np.ndarray:
        """
        Track a grid of points through the video.

        Returns:
            Array of shape (num_frames, num_points, 2) with point positions
        """
        # TODO: Implement actual CoTracker point tracking
        # This would:
        # 1. Initialize a grid of points in the first frame
        # 2. Track all points through subsequent frames
        # 3. Return point trajectories

        # Placeholder: return empty tracks
        num_frames = end_frame - start_frame
        num_points = self.grid_size * self.grid_size
        return np.zeros((num_frames, num_points, 2))

    async def _estimate_camera_motion(
        self,
        point_tracks: np.ndarray,
        fps: float,
        start_frame: int,
        end_frame: int,
    ) -> list[CameraMotion]:
        """
        Estimate camera motion from point trajectories.

        Uses homography estimation or fundamental matrix to derive
        camera motion parameters from point correspondences.
        """
        camera_motion = []
        num_frames = point_tracks.shape[0]

        for i in range(num_frames):
            frame_num = start_frame + (i * self.sample_rate)

            # TODO: Implement actual camera motion estimation
            # This would:
            # 1. Compute flow vectors between frames
            # 2. Fit homography to find global motion
            # 3. Decompose into pan/tilt/zoom/roll
            # 4. Estimate shake from residual motion

            # Placeholder values
            cm = CameraMotion(
                frame_number=frame_num,
                timestamp_seconds=frame_num / fps,
                pan=0.0,
                tilt=0.0,
                zoom=1.0,
                roll=0.0,
                shake=0.0,
            )
            camera_motion.append(cm)

        return camera_motion

    def smooth_motion(
        self,
        camera_motion: list[CameraMotion],
        window_size: int = 5,
    ) -> list[CameraMotion]:
        """
        Apply smoothing to camera motion data.

        Args:
            camera_motion: Raw camera motion data
            window_size: Smoothing window size

        Returns:
            Smoothed camera motion data
        """
        if len(camera_motion) < window_size:
            return camera_motion

        # Simple moving average smoothing
        smoothed = []

        for i, cm in enumerate(camera_motion):
            start = max(0, i - window_size // 2)
            end = min(len(camera_motion), i + window_size // 2 + 1)
            window = camera_motion[start:end]

            avg_pan = sum(c.pan for c in window) / len(window)
            avg_tilt = sum(c.tilt for c in window) / len(window)
            avg_zoom = sum(c.zoom for c in window) / len(window)
            avg_roll = sum(c.roll for c in window) / len(window)

            smoothed.append(CameraMotion(
                frame_number=cm.frame_number,
                timestamp_seconds=cm.timestamp_seconds,
                pan=avg_pan,
                tilt=avg_tilt,
                zoom=avg_zoom,
                roll=avg_roll,
                shake=cm.shake,  # Don't smooth shake
            ))

        return smoothed

    def detect_shake_intensity(
        self,
        camera_motion: list[CameraMotion],
    ) -> float:
        """
        Calculate overall camera shake intensity.

        Returns:
            Shake intensity from 0 (tripod) to 1 (handheld/chaotic)
        """
        if not camera_motion:
            return 0.0

        # Calculate variance of motion as proxy for shake
        pan_var = np.var([cm.pan for cm in camera_motion])
        tilt_var = np.var([cm.tilt for cm in camera_motion])

        # Normalize to 0-1 range
        total_var = pan_var + tilt_var
        normalized = min(1.0, total_var * 10)  # Scale factor

        return normalized
