requirements.txt
# Agentic Video - Python Dependencies

# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.1.0

# Web Server
fastapi>=0.109.0
uvicorn>=0.27.0
sse-starlette>=3.0.0  # For SSE streaming

# Async/HTTP
aiohttp>=3.9.0
httpx>=0.27.0
httpx-sse>=0.4.0  # For SSE client
tenacity>=8.0.0

# LLM
google-genai>=1.0.0
google-generativeai>=0.4.0  # Google Gemini SDK
langfuse>=2.0.0  # OSS LLM observability (self-hosted)

# Database
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
sqlalchemy>=2.0.0

# Cloud Storage
boto3>=1.34.0  # For R2 (S3-compatible)

# Video Processing
numpy>=1.24.0
scenedetect>=0.6.0  # Transition detection

# AI/ML (install separately for GPU)
# torch>=2.1.0
# torchvision>=0.16.0
# opencv-python>=4.8.0

# Utilities
asyncio-throttle>=1.0.0
typing-extensions>=4.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
